refactor: flatten runtime architecture and simplify provider config
Consolidate src/runtime/ utilities into src/gateway-*.ts modules (errors, config, retry, request). Replace GatewaySharedConfig (14 fields) with GatewayProviderSettings (8 fields) — remove logger, hooks, transport, autoRefreshToken, tokenCacheTTL, generateRequestId. Inline request execution logic. Update NestJS module, test fixtures, and imports. Delete src/runtime/ and src/types/ folders.
