--- src/__tests__/runtime-entry.spec.ts
@@ -1,15 +1,0 @@
-import { describe, it, expect } from 'vitest';
-import * as runtimeEntry from '../runtime';
-
-describe('runtime entry', () => {
-  it('exports core runtime primitives', () => {
-    expect(runtimeEntry.DEFAULT_BASE_URLS).toBeDefined();
-    expect(runtimeEntry.createFetchTransport).toBeDefined();
-    expect(runtimeEntry.AIGatewayError).toBeDefined();
-  });
-
-  it('does not export removed primitives', () => {
-    expect('SDKValidationError' in runtimeEntry).toBe(false);
-    expect('API_PATHS' in runtimeEntry).toBe(false);
-  });
-});
