+++ src/nestjs/constants.ts
@@ -1,4 +1,4 @@
-/** Injection token for the GatewaySharedConfig settings object. */
+/** Injection token for the GatewayProviderSettings object. */
 export const AI_GATEWAY_CLIENT = Symbol('AI_GATEWAY_CLIENT');
 
 /** Injection token for the module config options. */
