+++ src/nestjs/decorators.ts
@@ -2,13 +2,13 @@ import { Inject } from '@nestjs/common';
 import { AI_GATEWAY_CLIENT } from './constants';
 
 /**
- * Injects the configured `GatewaySharedConfig` settings into a constructor parameter or property.
+ * Injects the configured `GatewayProviderSettings` into a constructor parameter or property.
  *
  * @example
  * ```ts
  * @Injectable()
  * export class ChatService {
- *   constructor(@InjectAIGateway() private readonly config: GatewaySharedConfig) {}
+ *   constructor(@InjectAIGateway() private readonly config: GatewayProviderSettings) {}
  * }
  * ```
  */
