--- src/runtime/index.ts
@@ -1,12 +1,0 @@
-/**
- * Advanced runtime primitives used by the provider layer.
- * This entry owns the shared request pipeline.
- */
-
-export * from './config';
-export * from './errors';
-export { anySignal } from './abort';
-export { runRequest } from './request';
-export * from './retry';
-export { createFetchTransport } from './transport/fetch';
-export type { FetchTransportOptions } from './transport/fetch';
