--- src/types/index.ts
@@ -1,2 +1,0 @@
-export * from './codes';
-export type { GatewayApiErrorItem, GatewayApiErrorResponse, OpenAIErrorResponse } from './error-shapes';
